rename series to chartseries
